decouple PackageToRepositorySplitter
