+++ packages/Monorepo/src/Console/Command/SplitCommand.php
@@ -2,7 +2,6 @@
 
 namespace Symplify\Monorepo\Console\Command;
 
-use GitWrapper\GitWorkingCopy;
 use GitWrapper\GitWrapper;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -10,6 +9,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symplify\Monorepo\Configuration\ConfigurationGuard;
 use Symplify\Monorepo\Filesystem\Filesystem;
+use Symplify\Monorepo\PackageToRepositorySplitter;
 use Symplify\PackageBuilder\Console\Command\CommandNaming;
 use Symplify\PackageBuilder\Parameter\ParameterProvider;
 
@@ -40,18 +40,25 @@ final class SplitCommand extends Command
      */
     private $filesystem;
 
+    /**
+     * @var PackageToRepositorySplitter
+     */
+    private $packageToRepositorySplitter;
+
     public function __construct(
         ParameterProvider $parameterProvider,
         ConfigurationGuard $configurationGuard,
         GitWrapper $gitWrapper,
         SymfonyStyle $symfonyStyle,
-        Filesystem $filesystem
+        Filesystem $filesystem,
+        PackageToRepositorySplitter $packageToRepositorySplitter
     ) {
         $this->parameterProvider = $parameterProvider;
         $this->configurationGuard = $configurationGuard;
         $this->gitWrapper = $gitWrapper;
         $this->symfonyStyle = $symfonyStyle;
         $this->filesystem = $filesystem;
+        $this->packageToRepositorySplitter = $packageToRepositorySplitter;
 
         parent::__construct();
     }
@@ -77,7 +84,7 @@ final class SplitCommand extends Command
             $this->getSubsplitDirectory()
         ));
 
-        $this->splitDirectoriesToRepositories($gitWorkingCopy, $splitConfig);
+        $this->packageToRepositorySplitter->splitDirectoriesToRepositories($gitWorkingCopy, $splitConfig);
 
         $this->filesystem->deleteDirectory($this->getSubsplitDirectory());
         $this->symfonyStyle->success(sprintf('Directory "%s" cleaned', $this->getSubsplitDirectory()));
@@ -86,38 +93,5 @@ final class SplitCommand extends Command
     private function getSubsplitDirectory(): string
     {
         return getcwd() . '/.subsplit';
-    }
-
-    private function getMostRecentTag(GitWorkingCopy $gitWorkingCopy): string
-    {
-        $tags = $gitWorkingCopy->tag('-l');
-        $tagList = explode(PHP_EOL, trim($tags));
-
-        return (string) array_pop($tagList);
-    }
-
-    /**
-     * @todo own service
-     * @param mixed[] $splitConfig
-     */
-    private function splitDirectoriesToRepositories(GitWorkingCopy $gitWorkingCopy, array $splitConfig): void
-    {
-        $theMostRecentTag = $this->getMostRecentTag($gitWorkingCopy);
-
-        foreach ($splitConfig as $localSubdirectory => $remoteRepository) {
-            $this->splitLocalSubdirectoryToGitRepositoryWithTag(
-                $localSubdirectory,
-                $remoteRepository,
-                $theMostRecentTag
-            );
-        }
-    }
-
-    private function splitLocalSubdirectoryToGitRepositoryWithTag(
-        string $localSubdirectory,
-        string $remoteGitRepository,
-        ?string $theMostRecentTag = null
-    ): void {
-        // ...
     }
 }
